refactor(countdown): move to script setup
